add kohya to flux model weight output
